requirements.txt
# Core Dependencies
numpy>=1.24.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
pillow>=10.0.0
scipy>=1.11.0

# AI/ML Frameworks
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.0  # YOLO v8, v9, v10, v11
onnx>=1.15.0
onnxruntime>=1.16.0
tensorrt>=8.6.0  # For NVIDIA Jetson optimization
face-recognition>=1.3.0  # For people recognition
dlib>=19.24.0  # Required for face-recognition
albumentations>=1.4.0  # Advanced data augmentation for CNN training
timm>=0.9.0  # PyTorch Image Models (for additional backbones)
tqdm>=4.66.0  # Progress bars for downloads and training
optuna>=3.5.0  # Hyperparameter optimization framework

# Computer Vision & Tracking
filterpy>=1.4.5  # Kalman filters
scikit-image>=0.21.0
imutils>=0.5.4

# ROS 2 (install separately via apt)
# rclpy>=3.3.0
# sensor-msgs>=4.2.0
# geometry-msgs>=4.2.0
# nav-msgs>=4.2.0

# MAVLink & Drone Communication
pymavlink>=2.4.37
dronekit>=2.9.2

# Parrot Drones (ARSDK)
# Note: ARSDK requires separate installation from Parrot's repository
# olympe>=7.0.0  # Parrot SDK (install separately)

# DJI Drones
# dji-sdk>=4.0.0  # DJI SDK (requires separate installation)
# djitellopy>=2.5.0  # DJI Tello SDK

# ROS/ROS2 Support
# rclpy>=3.3.0  # ROS 2 Python client (install via apt)
# sensor-msgs>=4.2.0  # ROS sensor messages
# geometry-msgs>=4.2.0  # ROS geometry messages
# nav-msgs>=4.2.0  # ROS navigation messages
# mavros>=1.17.0  # MAVLink to ROS bridge (install via apt)

# WebRTC Support
aiortc>=1.5.0  # WebRTC implementation for Python
av>=10.0.0  # Audio/Video processing

# Video Streaming
opencv-python>=4.8.0  # Already included, supports RTSP/RTMP
ffmpeg-python>=0.2.0  # FFmpeg wrapper for video processing

# Network Protocols
aiohttp>=3.9.0  # Async HTTP client/server
websockets>=12.0  # Already included
pyserial>=3.5  # Serial communication
pyserial-asyncio>=0.6  # Async serial communication
pybluez>=0.23  # Bluetooth support (Linux only, requires system libs)

# Remote ID Support
# opendroneid>=1.0.0  # OpenDroneID implementation (if available)

# Sensor & Payload Support
pyserial>=3.5  # Already included above
pyserial-asyncio>=0.6  # Already included above

# Terrain Following & GPS-Denied Navigation
gdal>=3.7.0  # Geospatial data processing
rasterio>=1.3.0  # Raster I/O
shapely>=2.0.0  # Geometric operations
geopy>=2.4.0  # Geocoding and distance calculations

# SLAM & Visual Odometry (GPS-denied navigation)
# opencv-contrib-python>=4.8.0  # Already included, includes SLAM features
# ORB-SLAM3, RTAB-Map require separate installation

# Advanced Mission Planning
networkx>=3.2.0  # Graph algorithms for path planning
scikit-learn>=1.3.0  # Machine learning for adaptive planning

# Web Framework & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
aiofiles>=23.2.0
httpx>=0.25.0  # For Claude API integration

# Database & Storage
sqlalchemy>=2.0.23
alembic>=1.12.0
psycopg2-binary>=2.9.9  # PostgreSQL
redis>=5.0.1

# AWS Services
boto3>=1.34.0  # AWS SDK for Kinesis streaming

# Security & Encryption
cryptography>=41.0.7
pycryptodome>=3.19.0
pyjwt[crypto]>=2.8.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
loguru>=0.7.2
pyyaml>=6.0.1
click>=8.1.7
requests>=2.31.0  # HTTP requests (used in Dockerfile healthcheck and API calls)

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development Tools
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pylint>=3.0.3

# Performance Monitoring
psutil>=5.9.6
memory-profiler>=0.61.0

# Sensor Interfaces (if available)
# pyrealsense2>=2.54.0  # Intel RealSense
# flirpy>=0.2.0  # FLIR thermal cameras

